fix(desktop): harden app-owned read boundary
